Restructure polledtimeout, simplify namespaces
