+++ libraries/ESP8266WiFi/src/include/UdpContext.h
@@ -406,10 +406,10 @@ public:
     }
 
     bool sendTimeout(const ip_addr_t* addr, uint16_t port,
-                     esp8266::polledTimeout::oneShotFastMs::timeType timeoutMs)
+                     PolledTimeout::oneShotFastMs::timeType timeoutMs)
     {
         err_t err;
-        esp8266::polledTimeout::oneShotFastMs timeout(timeoutMs);
+        PolledTimeout::oneShotFastMs timeout(timeoutMs);
         while (((err = trySend(addr, port, /* keep buffer on error */true)) != ERR_OK) && !timeout)
             delay(0);
         if (err != ERR_OK)
