+++ libraries/Wire/examples/master_reader/master_reader.ino
@@ -23,7 +23,7 @@ void setup() {
 }
 
 void loop() {
-  using periodic = esp8266::polledTimeout::periodicMs;
+  using periodic = PolledTimeout::periodicMs;
   static periodic nextPing(1000);
 
   if (nextPing) {
